Hackthon/requirements.txt
pillow>=9.0.0
tkcalendar>=1.6.1
pynput>=1.7.6
PySimpleGUI>=4.60.0
PySide6>=6.0.0
pywin32>=228; platform_system == "Windows"
winshell>=0.6; platform_system == "Windows" 